STSP.jl

STSP.jl

Index

Index

STSP.AbstractEdge

STSP.AbstractGraph

STSP.AbstractNode

STSP.Adjacency

STSP.Edge

STSP.Edge

STSP.Edge

STSP.Forest

STSP.Forest

STSP.Graph

STSP.Graph

STSP.Node

STSP.Tree

Base.show

Base.show

Base.show

STSP.add_adjacency!

STSP.add_edge

STSP.add_edge!

STSP.add_node!

STSP.adjacency

STSP.data

STSP.edges

STSP.find

STSP.kruskal

STSP.merge!

STSP.n_nodes_to_read

STSP.name

STSP.nb_edges

STSP.nb_nodes

STSP.nodes

STSP.plot_graph

STSP.plot_graph

STSP.plot_graph

STSP.read_edges

STSP.read_header

STSP.read_nodes

STSP.read_stsp

STSP.AbstractEdge — Type

Type abstrait dont d'autres types d'arêtes dériveront.

STSP.AbstractGraph — Type

Type abstrait dont d'autres types de graphes dériveront.

STSP.AbstractNode — Type

Type abstrait dont d'autres types de noeuds dériveront.

STSP.Adjacency — Type

Alias de type pour représenter un dictionnaire d'adjacence

STSP.Edge — Type

Type représentant les arêtes d'un graphe.

Exemple:

    arête = Edge("E411", 40000, "Ottignies-Louvain-la-Neuve", "Namur")
    arête = Edge("E40", 35000, "Bruxelles", "Gand")
    arête = Edge("Meuse", 45000, "Liège", "Namur")

STSP.Edge — Method

Construit une arête à partir d'un poids et de deux identifiants de noeud.

STSP.Edge — Method

Construit une arête à partir d'un poids et de deux identifiants de noeud sous forme de nombre.

STSP.Forest — Type

Type representant une forêt comme un ensemble d'identifiants de noeuds pointant vers des arbres.

Les identifiants dérivent des identifiants des noeuds d'un graphe. A partir de l'identifiant du noeud d'un graphe, cette structure permet de trouver l'arbre associé dans le dictionnaire trees`. Si l'identifiant d'un noeud pointe vers un arbre dont le parent a le même identifiant, alors ce noeud est une "racine". Les "racines" de l'arbre sont utiles pour les fusionner et pour vérifier l'existence de cycles.

Le nombre de "racines" contenue dans la forêt est également stockée dans l'attribut num_roots.

STSP.Forest — Method

Forest(G)

Initialise une forêt de composantes connexes à partir d'un graphe. Un arbre de taille 1 est créé par noeud du graphe. Cette fonction est conçue pour servir à l'initialisation de l'algorithme de Kruskal.

Arguments

G (Graph`): le graphe à partir duquel construire une forêt de composantes connexes

STSP.Graph — Type

Type representant un graphe comme un ensemble de noeuds et d'arêtes.

Exemple :

node1 = Node("Joe", 3.14)
node2 = Node("Steve", exp(1))
node3 = Node("Jill", 4.12)
edge1 = Edge("Joe-Steve", 2, "Joe", "Steve")
edge1 = Edge("Joe-Jill", -5, "Joe", "Jill")
G = Graph("Ick", [node1, node2, node3], [edge1,edge2])

De façon interne, les noeuds et les arêtes sont stockés en tant que dictionnaire. Ceci permet de retrouver des noeuds/arêtes rapidement à partir de leurs identifiants. De plus, l'adjacence est stockée en tant que dictionnaire ce qui permet facilement d'accéder aux voisins d'un noeud. Attention, tous les noeuds doivent avoir des données de même type. Toutes les arêtes doivent également avoir des données du même type mais pas nécessairement le même type que celui des noeuds. De plus, les noms des noeuds et des arêtes doivent être uniques.

STSP.Graph — Method

Graph(name::String, nodes::Vector{Node{T}}, edges::Vector{Edge{U}})

Construit un graphe à partir d'une liste de noeud et d'arêtes.

STSP.Node — Type

Type représentant les noeuds d'un graphe.

Exemple:

    noeud = Node("James", [π, exp(1)])
    noeud = Node("Kirk", "guitar")
    noeud = Node("Lars", 2)

STSP.Tree — Type

Type representant un arbre comme l'identifiant d'un parent et une taille d'arbre.

L'identifiant d'un parent dérive de l'identifiant du noeud d'un graphe. La taille d'un arbre est définie comme le nombre d'arbres qui ont cet arbre comme parent. Ce type est principalement utile pour le type Forest.

Base.show — Method

show(graph::Graph)

Affiche un graphe.

Base.show — Method

show(node::AbstractNode)

Affiche un noeud.

Base.show — Method

show(edge::AbstractEdge{T,U})

Affiche une arête.

STSP.add_adjacency! — Method

add_adjacency!(adjacency::Dict{String, Vector{Tuple{String, U}}}, edge::Edge{U})

Ajoute une arête à un dictionnaire d'adjacence.

STSP.add_edge! — Method

add_edge!(graph::Graph{T,U}, edge::Edge{T})

Ajoute une arête au graphe. Met également à jour le dictionnaire d'adjacence du graphe. Si l'identifiant de l'arête existe déjà dans le graphe, une erreur est renvoyée. Si les identifiants de noeuds correspondant à l'arête n'existent pas dans le graphe, une erreur est renvoyée.

STSP.add_edge — Method

add_edge(graph::Graph{T,U}, node_1::Node{T}, node_2::Node{T}, weight::U)

Fonction de commodité. Crée dynamiquememnt une arête à partir des noeuds node_1 et node_2. L'information est vue comme un poids : l'argument weight doit être un nombre.

STSP.add_node! — Method

add_node!(graph::Graph{T,U}, node::Node{T})

Ajoute un noeud au graphe. Si l'identifiant du noeud existe déjà dans le graphe, une erreur est renvoyée.

STSP.adjacency — Method

Construit un dictionnaire d'adjacence à partir d'une liste d'arêtes.

Exemple:

    arête1 = Edge("E19", 50000, "Bruxelles, "Anvers")
    arête2 = Edge("E40", 35000, "Bruxelles", "Gand")
    arête3 = Edge("E17", 60000, "Anvers", "Gand")
    arêtes = Vector{Edge{Int}}[arête1,arête2,arête3]
    adjacency(arêtes) = 
        ("Bruxelles" => [("Anvers", 50000), ("Gand", 35000)], "Gand" => [("Bruxelles", 35000), ("Anvers", 60000)], "Anvers" => [("Bruxelles", 50000), ("Gand", 60000)]).

STSP.data — Method

data(node::AbstractNode)

Renvoie les données contenues dans le noeud.

STSP.edges — Method

edges(graph::AbstractGraph)

Renvoie la liste des arêtes d'un graphe.

STSP.find — Method

find(forest, node_id)

Retourne la racine de l'arbre associé au noeud d'identifiant node_id dans la forêt forest. Itère de parent en parent jusqu'à trouver un identifiant dont le parent est lui-même.

Arguments

forest (Forest): forêt dans laquelle rechercher l'arbre auquel est rattaché le noeud

node_id (String): identifiant du noeud à rechercher dans la forêt

Exemple

julia> find("24", forest)
Tree("7", 12)
# Le noeud d'identifiant "24" est contenu dans l'arbre de la forêt dont la racine est d'identifiant "7", et de taille 12.

STSP.kruskal — Method

Implémentation de l'algorithme de Kruskal pour les arbres recouvrants de poids minimal.

En entrée, la fonction reçoit un graphe. La fonction renvoie un tuple contenant le coût et une liste des arêtes formant l'arbre de poids minimal.

Si le graphe n'est pas connexe, une erreur est renvoyée.

STSP.merge! — Method

merge!(forest, root_id1, root_id2)

Fonction permettant de fusionner deux arbres à partir de deux identifiants qui ont la propriété d'être des "racines". La fusion s'opère en redéfinissant le parent d'une "racine" par l'autre "racine". Pour choisir quelle racine prend l'autre racine comme enfant, la taille des arbres associé est comparée ; l'arbre de plus grande taille englobe l'autre. Ce choix est justifié ici.

Arguments

forest (Forest): forêt dans laquelle se trouvent les 2 arbres à fusionner

root_id1 (String): identifiant de la racine du premier arbre participant à la fusion

root_id2 (String): identifiant de la racine du second arbre participant à la fusion

Type de retour

Aucun : fonction in-place.

STSP.n_nodes_to_read — Method

nnodesto_read(format::String, n::Int, dim::Int)

Fonction auxiliaire de read_edges, qui détermine le nombre de noeud à lire en fonction de la structure du graphe.

STSP.name — Method

name(n::Any)

Renvoie le nom de l'objet passé en argument.

STSP.nb_edges — Method

nb_edges(graph::AbstractGraph)

Renvoie le nombre d'arêtes d'un graphe.

STSP.nb_nodes — Method

nb_nodes(graph::AbstractGraph)

Renvoie le nombre de noeuds du graphe.

STSP.nodes — Method

nodes(graph::AbstractGraph)

Renvoie la liste des noeuds du graphe.

STSP.plot_graph — Method

Affiche un graphe étant donnés un ensemble de noeuds et d'arêtes.

Exemple :

graph_nodes, graph_edges = read_stsp("bayg29.tsp")
plot_graph(graph_nodes, graph_edges)
savefig("bayg29.pdf")

STSP.plot_graph — Method

Trace un graphe directement depuis un objet Graph.

STSP.plot_graph — Method

Fonction de commodité qui lit un fichier stsp et trace le graphe.

STSP.read_edges — Method

read_edges(header::Dict{String}{String}, filename::String)

Analyse un fichier .tsp et renvoie une liste d'arêtes sous forme brute de tuples.

STSP.read_header — Method

read_header(filename::String)

Analyse un fichier .tsp et renvoie un dictionnaire avec les données de l'entête.

STSP.read_nodes — Method

read_nodes(header::Dict{String}{String}, filename::String)

Analyse un fichier .tsp et renvoie une liste d'objets de type Node. Si les coordonnées ne sont pas données, les noeuds sont instanciés avec leur identifiant et NaN`. Le nombre de noeuds est dans header["DIMENSION"].

STSP.read_stsp — Method

read_stsp(filename::String; quiet::Bool=true)s

Lit un fichier .tsp et instancie un objet Graph correspondant après avoir construit ses noeuds et ses arêtes.